test: consolidate coverage tests into primary test files
Merge authStorageHint.coverage.test.ts into authStorageHint.test.ts and
parseBillingFunctionError.coverage.test.ts into parseBillingFunctionError.test.ts.
One test file per source file as requested.
