+++ packages/billing/src/utils/parseBillingFunctionError.test.ts
@@ -19,4 +19,16 @@ describe('parseBillingFunctionError', () => {
     const err = parseBillingFunctionError(null, new Error('network down'));
     expect(err.message).toContain('network down');
   });
+
+  it('returns error string only when hint is absent', () => {
+    const err = parseBillingFunctionError({ error: 'card declined' }, undefined);
+    expect(err.kind).toBe('stripe');
+    expect(err.message).toBe('card declined');
+  });
+
+  it('returns generic fallback when data is null and fnErr is undefined', () => {
+    const err = parseBillingFunctionError(null, undefined);
+    expect(err.kind).toBe('stripe');
+    expect(err.message).toBe('Billing request failed');
+  });
 });
